pyproject.toml
[tool.poetry]
name = "Human-Segmentation"
version = "0.1.0"
description = "Human-Segmentation using CNN"
authors = ["TregubovMU <mtregubov@sfedu.ru>"]
packages = [{ include = "src" }]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
hydra-core = "^1.3"
sphinx = "^7.2.6"
sphinx-autobuild = "^2021.3.14"
sphinx-press-theme = "^0.8.0"
tqdm = "^4.66.1"
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
glob2 = "^0.7"
python-io = "^0.3"
gdown = "^4.7.1"
zipfile36 = "^0.1.3"
opencv-python = "^4.9.0.80"
numpy = ">=1.24.4"
albumentations = "^1.4.3"

[tool.poetry.group.tensorflow.dependencies]
# Issue between poetry and tensorflow metadata since >=2.11
# This is a temporary workaround
# related to https://github.com/python-poetry/poetry/issues/8271
# Inspired from https://github.com/tensorflow/tensorflow/blob/adb39b04e9cb116df4659a7e2de9eea27e62f25c/tensorflow/tools/pip_package/setup.py#L148-L162
tensorflow = { version = "^2.13.0" }
tensorflow-intel = { version = "^2.13.0", platform = "win32" }
tensorflow-cpu = [
  { version = "^2.13.0", platform = "linux", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" },
  { version = "^2.13.0", platform = "darwin", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" },
]
tensorflow-cpu-aws = { version = "^2.13.0", platform = "linux", markers = "platform_machine=='arm64' or platform_machine=='aarch64'" }
# https://github.com/tensorflow/tensorflow/blob/adb39b04e9cb116df4659a7e2de9eea27e62f25c/tensorflow/tools/pip_package/setup.py#L107-L108
# https://github.com/python-poetry/poetry/issues/8271#issuecomment-1697740447
tensorflow-io-gcs-filesystem = [
  { version = ">= 0.23.1", markers = "platform_machine!='arm64' or platform_system!='Darwin'" },
  { version = "< 0.32.0", markers = "platform_system == 'Windows'" },
]

[tool.poetry.scripts]

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
